requirements.txt
click==6.7
Flask==0.12.2
Flask-SQLAlchemy==2.2
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==0.23
psycopg2==2.7.1
PyMySQL==0.7.9
python-dotenv==0.6.1
SQLAlchemy==1.1.11
waitress==1.0.2
Werkzeug==0.12.2
